feature: test for collision

+++ utility.go
@@ -1,13 +1,13 @@
 package main
 
-func Min(a, b int) int {
+func Min(a, b float32) float32 {
 	if a < b {
 		return a
 	}
 	return b
 }
 
-func Max(a, b int) int {
+func Max(a, b float32) float32 {
 	if a > b {
 		return a
 	}
